Move Handler observers from handler to handler.observer
