+++ api/metrics-api-jaxrs/src/main/java/org/hawkular/metrics/api/jaxrs/handler/observer/ResultSetObserver.java
@@ -14,7 +14,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.metrics.api.jaxrs;
+package org.hawkular.metrics.api.jaxrs.handler.observer;
+
+import org.hawkular.metrics.api.jaxrs.ApiError;
 
 import com.datastax.driver.core.ResultSet;
